requirements.txt
torch>=1.13.0
torchvision>=0.14.0
timm>=0.6.13
numpy>=1.21.0
Pillow>=9.0.0
tqdm>=4.64.0
plotly>=5.9.0
